pyproject.toml
[project]
name = "research-template"
version = "0.1.0"
description = "Add a short description here"
authors = [{ name = "tubo213", email = "hoge@gmail.com" }]
dependencies = [
    "numpy>=1.24.4",
    "matplotlib>=3.7.2",
    "tqdm>=4.65.0",
    "torch==2.1.0",
    "joblib>=1.3.1",
    "pyyaml>=6.0.1",
    "types-PyYAML>=6.0.12.11",
    "hydra-core>=1.3.2",
    "pytorch_pfn_extras>=0.7.1",
    "scipy>=1.10.1",
    "pandas>=2.0.3",
    "pytorch-lightning>=2.0.6",
    "polars>=0.18.9",
    "scikit-learn>=1.3.0",
    "streamlit>=1.25.0",
    "kaggle~=1.5.16",
    "jupyter-black~=0.3.4",
    "japanize_matplotlib~=1.1.3",
    "seaborn~=0.12.2",
    "transformers~=4.33.2",
    "setuptools~=68.2.2",
    "wandb~=0.15.11",
    "torchaudio>=2.0.2",
    "librosa>=0.10.1",
    "numba>=0.58.0",
    "torchvision>=0.15.2",
    "segmentation-models-pytorch>=0.3.3",
    "rich>=13.6.0",
    "types-tqdm>=4.66.0.2",
    "optuna>=3.4.0",
    "plotly>=5.18.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[[tool.rye.sources]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.4.1",
    "jupyter>=1.0.0",
    "ruff~=0.0.291",
    "pytest>=7.4.2",
]

[tool.hatch.metadata]
allow-direct-references = true
